sio: set diag in case errno == EADDRINUSE
In iproto.cc and in coio.cc we raise diag in case when
listen fails, but if errno == EADDINUSE diag is not set,
which leads to assetion. Also there is no reason to ignore
EADDRINUSE for `listen` (for `bind` we check, that socket
not used and if this is UNIX socket close it).