test: flaky replication/long_row_timeout.test.lua
On heavy loaded hosts either slow machines like VMware found the
following issues:

  [010] --- replication/long_row_timeout.result   Fri May  8 08:56:08 2020
  [010] +++ var/rejects/replication/long_row_timeout.reject       Mon Jun 21 04:39:08 2021
  [010] @@ -23,7 +23,7 @@
  [010]  ...
  [010]  box.info.replication[2].downstream.status
  [010]  ---
  [010] -- follow
  [010] +- stopped
  [010]  ...
  [010]  -- make applier incapable of reading rows in one go, so that it
  [010]  -- yields a couple of times.
  [010]

  [026] --- replication/long_row_timeout.result	Fri Sep 25 19:08:46 2020
  [026] +++ replication/long_row_timeout.reject	Fri May  8 08:23:39 2020
  [026] @@ -80,7 +80,7 @@
  [026]  ...
  [026]  ok
  [026]  ---
  [026] -- true
  [026] +- false
  [026]  ...
  [026]  s:drop()
  [026]  ---
  [026]

It happened because replication downstream status check occurred too
early, when it was only in 'stopped' state. This situation happens
when replica done its initial join, but not reached subscription yet.
To give the replication status check routine ability to reach the
needed 'follow' state, it need for it using test_run:wait_downstream()
routine.

The second issue happened because replica had not enought time to
check the status. To avoid of it replication_timeout must be tunned
for slow test hosts. Found that on FreeBSD this timeout need to be
multiplied in 5 times, it makes test to run 0.2 sec on such hosts.

Closes #4351