sql: always show value in type mismatch error
Currently, in most cases of type mismatch errors, a value that cannot be
cast is printed in the description. An exception is the description of
the type mismatch error in the comparison. There, instead of the value,
the type of the value is printed. This patch removes this exception, so
now the value will always be printed in the type mismatch error
description.

Follow up #4356